Rearrange dbus exports
