--- crystalfontz/dbus/__init__.py
@@ -1,13 +1,0 @@
-try:
-    from crystalfontz.dbus.client import DbusClient
-    from crystalfontz.dbus.interface import DBUS_NAME, DbusInterface
-except ImportError:
-    DbusClient = None
-    DBUS_NAME = None
-    DbusInterface = None
-
-__all__ = [
-    "DbusClient",
-    "DbusInterface",
-    "DBUS_NAME",
-]
