+++ crystalfontz/dbus/client/cli.py
@@ -8,8 +8,7 @@ import shlex
 import shutil
 import subprocess
 import sys
-from typing import Any, cast, List, Optional, Self, Tuple
-from unittest.mock import Mock
+from typing import cast, List, Optional, Tuple
 
 import click
 from sdbus import (  # pyright: ignore [reportMissingModuleSource]
@@ -36,11 +35,10 @@ from crystalfontz.cli import (
     OutputMode,
     WATCHDOG_SETTING,
 )
-from crystalfontz.config import Config
+from crystalfontz.dbus.client import DbusClient
 from crystalfontz.dbus.config import StagedConfig
 from crystalfontz.dbus.domain import (
     BytesM,
-    ConfigM,
     CursorStyleM,
     DowDeviceInformationM,
     DowTransactionResultM,
@@ -57,7 +55,6 @@ from crystalfontz.dbus.domain import (
     VersionsM,
 )
 from crystalfontz.dbus.error import handle_dbus_error
-from crystalfontz.dbus.interface import DBUS_NAME, DbusInterface
 from crystalfontz.gpio import GpioDriveMode, GpioFunction
 from crystalfontz.lcd import LcdRegister
 from crystalfontz.temperature import (
@@ -67,31 +64,6 @@ from crystalfontz.temperature import (
 )
 
 logger = logging.getLogger(__name__)
-
-
-class DbusClient(DbusInterface):
-    """
-    A DBus client for the Crystalfontz device.
-    """
-
-    def __init__(self: Self, bus: Optional[SdBus] = None) -> None:
-        client = Mock(name="client", side_effect=NotImplementedError("client"))
-        self.subscribe = Mock(name="client.subscribe")
-        super().__init__(client)
-
-        cast(Any, self)._proxify(DBUS_NAME, "/", bus=bus)
-
-    async def staged_config(self: Self) -> StagedConfig:
-        """
-        Fetch the state of staged configuration changes.
-        """
-
-        active_config: Config = ConfigM.unpack(await self.config)
-
-        return StagedConfig(
-            target_config=Config.from_file(active_config.file),
-            active_config=active_config,
-        )
 
 
 @dataclass
