+++ docs/api/crystalfontz.dbus.client.md
@@ -1,0 +1,3 @@
+# crystalfontz.dbus.client
+
+::: crystalfontz.dbus.client
