+++ docs/api/crystalfontz.dbus.interface.md
@@ -1,0 +1,3 @@
+# crystalfontz.dbus.interface
+
+::: crystalfontz.dbus.interface
