--- docs/api/crystalfontz.md
@@ -1,3 +1,0 @@
-# crystalfontz
-
-::: crystalfontz
