+++ pyproject.toml
@@ -42,7 +42,7 @@ dbus = [
 [project.scripts]
 crystalfontz = "crystalfontz.cli:main"
 crystalfontzd = "crystalfontz.dbus.service:main"
-crystalfontzctl = "crystalfontz.dbus.client:main"
+crystalfontzctl = "crystalfontz.dbus.cli:main"
 
 [dependency-groups]
 dev = [
